requirements.txt
# pip3 install -U -r requirements.txt
# conda install numpy opencv matplotlib mpld3
numpy
pyfunctional
# matplotlib
# mpld3
